Extract ShellIntegrationSweep helper + gate uninstall on profile presence
Refactor: replace three duplicated copies of the profile-snapshot +
install loop (FreOverlay, _InitShellIntegration, _ReconcileShellIntegration)
with a single ShellIntegrationSweep header. Each call site now reduces
to:

    const auto wslDistros    = SnapshotWslDistroNames(_settings);
    const auto shellPresence = SnapshotShellPresence(_settings);
    co_await winrt::resume_background();
    const auto results = RunInstall(shellPresence, wslDistros);   // or RunUninstall

Behavior change: profile-gate the uninstall path the same way install
is gated. Toggling shell integration off no longer touches every
supported shell unconditionally — it only sweeps shells the user
actually has a profile for (and WSL distros bounded by the snapshot,
as before). Trade-off: if the user installs for shell X, deletes the
X profile, then toggles off, the X block in HOME survives. This is
symmetric with install-time behavior (we never wrote for shells
without a profile either) and avoids creating .bak.* backups for
shells the user does not use. The next reconcile after re-adding the
X profile sweeps it.

PowerShell + Windows PowerShell already share BuildFlavor, ScriptContent,
IsOrphanBodyLine, BuildBlock, and FindLegacyDotSource via
PowerShellShellIntegration.h. WSL bash already reuses the entire Bash
flavor via Wsl::Install. No duplication remains across the per-flavor
modules.

Net diff: -137 lines (three copies collapse to one shared helper).
