+++ src/cascadia/TerminalApp/AppActionHandlers.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) Microsoft Corporation.
+﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
 #include "pch.h"
@@ -9,6 +9,7 @@
 #include "AgentPaneLog.h"
 #include "ScratchpadContent.h"
 #include "../inc/ShellIntegration.h"
+#include "ShellIntegrationSweep.h"
 #include "../WinRTUtils/inc/WtExeUtils.h"
 #include "../../types/inc/utils.hpp"
 #include "../TerminalSettingsAppAdapterLib/TerminalSettings.h"
@@ -35,148 +36,6 @@ namespace winrt
 
 namespace winrt::TerminalApp::implementation
 {
-    namespace
-    {
-        // Parse a WSL distro name out of a profile generated by
-        // WslDistroGenerator. Prefer the `-d <distName>` token in
-        // the commandline (the value WSL itself consumes) over
-        // profile.Name(), which is a free-text display field that
-        // the user can edit. Returns an empty string when the
-        // commandline parse fails (fail-closed): we do NOT fall
-        // back to profile.Name() because using user-editable display
-        // text as a distro identifier could write integration to the
-        // wrong distro or produce a wsl.exe spawn failure. Callers
-        // (e.g. _SnapshotWslDistroNames) drop empty results.
-        //
-        // (Duplicated from FreOverlay.cpp — small + symmetric.
-        // Promote to a shared helper if a third call site arrives.)
-        inline std::wstring _ExtractWslDistroName(const Profile& profile)
-        {
-            const std::wstring cmd{ profile.Commandline() };
-            constexpr std::wstring_view shortFlag{ L" -d " };
-            constexpr std::wstring_view longFlag{ L" --distribution " };
-            for (const auto flag : { shortFlag, longFlag })
-            {
-                const auto pos = cmd.find(flag);
-                if (pos == std::wstring::npos)
-                {
-                    continue;
-                }
-                const auto tokStart = pos + flag.size();
-                auto tokEnd = tokStart;
-                while (tokEnd < cmd.size() && cmd[tokEnd] != L' ' && cmd[tokEnd] != L'\t')
-                {
-                    ++tokEnd;
-                }
-                if (tokEnd > tokStart)
-                {
-                    return cmd.substr(tokStart, tokEnd - tokStart);
-                }
-            }
-            // Commandline parse failed. Do NOT fall back to profile.Name() —
-            // that's user-editable display text (see
-            // doc/cascadia/profiles.schema.json), so using it as a distro
-            // identifier would write integration to the wrong distro or
-            // produce a wsl.exe spawn failure. Return empty; the caller
-            // (_SnapshotWslDistroNames) drops empty results.
-            return {};
-        }
-
-        // Snapshot the user's WSL distros from the live settings model.
-        // MUST be called on the UI thread — _settings.AllProfiles()
-        // returns the live observable vector, and iterating it
-        // concurrently with a settings reload is a use-after-mutate.
-        inline std::vector<std::wstring> _SnapshotWslDistroNames(const CascadiaSettings& settings)
-        {
-            std::vector<std::wstring> out;
-            if (!settings)
-            {
-                return out;
-            }
-            // Multiple WT profiles can point at the same WSL distro
-            // (a user may duplicate "Ubuntu" with different startup
-            // commands, fonts, etc.). De-duplicate so each distro is
-            // touched exactly once per install/uninstall sweep — the
-            // per-distro path is also expensive (wsl.exe spawn, UNC
-            // mount, $HOME probe), so repeating it is doubly bad.
-            std::set<std::wstring> seen;
-            for (const auto& profile : settings.AllProfiles())
-            {
-                if (profile.Source() == L"Windows.Terminal.Wsl")
-                {
-                    auto name = _ExtractWslDistroName(profile);
-                    if (name.empty())
-                    {
-                        // _ExtractWslDistroName couldn't parse the
-                        // commandline. Skipping rather than attempting
-                        // an InstallWslBash("") that's guaranteed to
-                        // fail (and produces a noisy error in the FRE
-                        // / Settings UI). User's malformed WSL profile
-                        // simply doesn't get integration; the rest do.
-                        continue;
-                    }
-                    if (seen.insert(name).second)
-                    {
-                        out.emplace_back(std::move(name));
-                    }
-                }
-            }
-            return out;
-        }
-
-        // Snapshot which non-WSL shells the user has at least one
-        // profile for. MUST be called on the UI thread for the same
-        // reason as _SnapshotWslDistroNames — _settings.AllProfiles() is
-        // an observable live vector.
-        //
-        // Used to gate Install: writing a shell-integration block for
-        // a shell the user has no profile for would pollute their HOME
-        // with a file they will never source. WSL is not part of this
-        // snapshot because the per-distro list above is already the
-        // gate for WSL bash.
-        struct ShellPresence
-        {
-            bool pwsh{ false };
-            bool windowsPowerShell{ false };
-            bool bash{ false };
-        };
-        inline ShellPresence _SnapshotShellPresence(const CascadiaSettings& settings)
-        {
-            ShellPresence out{};
-            if (!settings)
-            {
-                return out;
-            }
-            namespace SI = ::Microsoft::Terminal::ShellIntegration;
-            for (const auto& profile : settings.AllProfiles())
-            {
-                // Hold the hstring values in named locals so their lifetime
-                // extends through the wstring_view uses below — wstring_view
-                // is non-owning and a temporary hstring would dangle.
-                const auto src = profile.Source();
-                const auto cmd = profile.Commandline();
-                const std::wstring_view srcSv{ src };
-                const std::wstring_view cmdSv{ cmd };
-                if (!out.pwsh && SI::ProfileMatchesShell(SI::Target::Pwsh, srcSv, cmdSv))
-                {
-                    out.pwsh = true;
-                }
-                if (!out.windowsPowerShell && SI::ProfileMatchesShell(SI::Target::WindowsPowerShell, srcSv, cmdSv))
-                {
-                    out.windowsPowerShell = true;
-                }
-                if (!out.bash && SI::ProfileMatchesShell(SI::Target::Bash, srcSv, cmdSv))
-                {
-                    out.bash = true;
-                }
-                if (out.pwsh && out.windowsPowerShell && out.bash)
-                {
-                    break;
-                }
-            }
-            return out;
-        }
-    }
     TermControl TerminalPage::_senderOrActiveControl(const IInspectable& sender)
     {
         if (sender)
@@ -2117,8 +1976,8 @@ namespace winrt::TerminalApp::implementation
         // profiles for, on the UI thread BEFORE we go background.
         // _settings.AllProfiles() is an observable vector; iterating it
         // concurrently with a settings reload would be unsafe.
-        const auto wslDistros = _SnapshotWslDistroNames(_settings);
-        const auto shellPresence = _SnapshotShellPresence(_settings);
+        const auto wslDistros = ShellIntegrationSweep::SnapshotWslDistroNames(_settings);
+        const auto shellPresence = ShellIntegrationSweep::SnapshotShellPresence(_settings);
 
         co_await winrt::resume_background();
 
@@ -2131,8 +1990,6 @@ namespace winrt::TerminalApp::implementation
         {
             co_return;
         }
-
-        namespace SI = ::Microsoft::Terminal::ShellIntegration;
 
         bool desiredAtRun = true;
         bool pwshAlready = false;
@@ -2150,47 +2007,16 @@ namespace winrt::TerminalApp::implementation
             desiredAtRun = _shellIntegrationDesiredEnabled.load(std::memory_order_acquire);
             if (desiredAtRun)
             {
-                // Profile-gated install: only write the integration
-                // block for shells the user actually has a profile for.
-                // A user with only "Developer PowerShell for VS" (which
-                // uses Windows PowerShell) and no pwsh profile should
-                // not get a pwsh profile block written. The success
-                // verdict treats a skipped shell as "already configured"
-                // for the UI's all-installed/any-failure summary so we
-                // don't flag the user's missing-shell as a failure.
-                SI::InstallResult pwshResult{};
-                SI::InstallResult wpResult{};
-                if (shellPresence.pwsh)
-                {
-                    pwshResult = SI::InstallForTarget(SI::Target::Pwsh);
-                }
-                else
-                {
-                    pwshResult.success = true;
-                    pwshResult.alreadyInstalled = true;
-                }
-                if (shellPresence.windowsPowerShell)
-                {
-                    wpResult = SI::InstallForTarget(SI::Target::WindowsPowerShell);
-                }
-                else
-                {
-                    wpResult.success = true;
-                    wpResult.alreadyInstalled = true;
-                }
-                // Bash + WSL are best-effort here too — their failures
-                // do not gate the "all installed" / "any failure" UI
-                // verdict. Users without Git Bash or without a (running)
-                // WSL distro should not see error dialogs. Profile-gated
-                // for the same reason as pwsh/Windows PowerShell.
-                if (shellPresence.bash)
-                {
-                    (void)SI::InstallForTarget(SI::Target::Bash);
-                }
-                for (const auto& distName : wslDistros)
-                {
-                    (void)SI::InstallWslBash(distName);
-                }
+                // Profile-gated install: RunInstall only touches shells
+                // the user actually has a profile for. A user with only
+                // "Developer PowerShell for VS" (Windows PowerShell) and
+                // no pwsh profile should not get a pwsh block written.
+                // Skipped shells are reported as success-already-installed
+                // so the all-installed / any-failure UI verdict below
+                // doesn't flag a missing shell as a failure.
+                const auto results = ShellIntegrationSweep::RunInstall(shellPresence, wslDistros);
+                const auto& pwshResult = results.pwsh;
+                const auto& wpResult = results.windowsPowerShell;
                 pwshAlready = pwshResult.alreadyInstalled;
                 wpAlready = wpResult.alreadyInstalled;
                 pwshOk = pwshResult.success;
@@ -2285,8 +2111,8 @@ namespace winrt::TerminalApp::implementation
         // thread BEFORE going background. _settings.AllProfiles() is
         // an observable vector and must not be iterated concurrently
         // with a settings reload.
-        const auto wslDistros = _SnapshotWslDistroNames(_settings);
-        const auto shellPresence = _SnapshotShellPresence(_settings);
+        const auto wslDistros = ShellIntegrationSweep::SnapshotWslDistroNames(_settings);
+        const auto shellPresence = ShellIntegrationSweep::SnapshotShellPresence(_settings);
 
         co_await winrt::resume_background();
         auto self = weak.get();
@@ -2305,55 +2131,32 @@ namespace winrt::TerminalApp::implementation
         std::lock_guard<std::mutex> guard{ _shellIntegrationReconcileMutex };
         const bool enabled = _shellIntegrationDesiredEnabled.load(std::memory_order_acquire);
 
-        namespace SI = ::Microsoft::Terminal::ShellIntegration;
         if (enabled)
         {
-            // Profile-gated install: only write to shells the user has
-            // a profile for. A user keeping only "Developer PowerShell
+            // Profile-gated install: only touch shells the user has a
+            // profile for. A user keeping only "Developer PowerShell
             // for VS" (which uses Windows PowerShell) and no pwsh
             // profile must not get pwsh integration written.
-            if (shellPresence.pwsh)
-            {
-                SI::InstallForTarget(SI::Target::Pwsh);
-            }
-            if (shellPresence.windowsPowerShell)
-            {
-                SI::InstallForTarget(SI::Target::WindowsPowerShell);
-            }
-            if (shellPresence.bash)
-            {
-                SI::InstallForTarget(SI::Target::Bash);
-            }
-            for (const auto& distName : wslDistros)
-            {
-                SI::InstallWslBash(distName);
-            }
+            (void)ShellIntegrationSweep::RunInstall(shellPresence, wslDistros);
         }
         else
         {
-            // Uninstall sweep policy:
-            //   * Non-WSL targets (Pwsh, WindowsPowerShell, Bash) are
-            //     NOT gated on shellPresence so a profile the user
-            //     deleted after a prior install is still cleaned up.
-            //     Uninstall is idempotent when no block is present.
-            //   * WSL is bounded by `wslDistros` (the CURRENT profile
-            //     list). If the user deleted a WSL distro profile
-            //     after install, the integration block inside that
-            //     distro's filesystem will NOT be cleaned up here.
-            //     That's intentional: WT profile deletion ≠ WSL distro
-            //     removal (the user may still use the distro via
-            //     `wsl.exe` directly), and we'd need to track
-            //     previously-installed distros across settings reloads
-            //     to do better — adds complexity for a rare edge case.
-            //     If the user re-adds the WT profile later, the next
-            //     reconcile/uninstall will clean it.
-            SI::UninstallForTarget(SI::Target::Pwsh);
-            SI::UninstallForTarget(SI::Target::WindowsPowerShell);
-            SI::UninstallForTarget(SI::Target::Bash);
-            for (const auto& distName : wslDistros)
-            {
-                SI::UninstallWslBash(distName);
-            }
+            // Profile-gated uninstall: symmetric with install — we
+            // only clean up shells the user currently has a profile
+            // for. Trade-off: if the user installed for shell X,
+            // deleted the X profile, then toggled off, the X block
+            // in their HOME survives. This matches install-time
+            // policy and avoids touching shells the user does not
+            // use (which would write `.bak.*` for nothing). The
+            // next reconcile after re-adding the X profile sweeps
+            // it.
+            //
+            // WSL is similarly bounded by `wslDistros`: WT profile
+            // deletion != WSL distro removal (the user may still
+            // use the distro via `wsl.exe` directly), and tracking
+            // previously-installed distros across settings reloads
+            // would add complexity for a rare edge case.
+            ShellIntegrationSweep::RunUninstall(shellPresence, wslDistros);
         }
     }
