+++ src/cascadia/TerminalApp/FreOverlay.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) Microsoft Corporation.
+﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
 #include "pch.h"
@@ -11,6 +11,7 @@
 #include "../inc/ShellIntegration.h"
 #include "../inc/RtlHelper.h"
 #include "AgentPaneLog.h"
+#include "ShellIntegrationSweep.h"
 
 #include <winrt/Windows.UI.Xaml.Documents.h>
 
@@ -22,154 +23,6 @@ namespace Automation = winrt::Windows::UI::Xaml::Automation;
 
 namespace winrt::TerminalApp::implementation
 {
-    namespace
-    {
-        // Parse a WSL distro name out of a profile generated by
-        // WslDistroGenerator. We prefer the `-d <distName>` token in
-        // the commandline (the value WSL itself consumes) over
-        // `profile.Name()` — Name() defaults to distName but is a
-        // free-text display field the user can rename; the
-        // commandline survives a rename. Returns an empty string when
-        // the commandline parse fails (fail-closed): we do NOT fall
-        // back to `profile.Name()` because using user-editable display
-        // text as a distro identifier could write integration to the
-        // wrong distro or produce a wsl.exe spawn failure. Callers
-        // (e.g. _SnapshotWslDistroNames) drop empty results.
-        //
-        // Caller is responsible for IsSafeWslDistroName validation
-        // before passing the result to wsl.exe.
-        inline std::wstring _ExtractWslDistroName(const winrt::Microsoft::Terminal::Settings::Model::Profile& profile)
-        {
-            const std::wstring cmd{ profile.Commandline() };
-            // Look for ` -d <token>` or ` --distribution <token>`.
-            // Whitespace-bounded so we don't match `--default-d` etc.
-            constexpr std::wstring_view shortFlag{ L" -d " };
-            constexpr std::wstring_view longFlag{ L" --distribution " };
-            for (const auto flag : { shortFlag, longFlag })
-            {
-                const auto pos = cmd.find(flag);
-                if (pos == std::wstring::npos)
-                {
-                    continue;
-                }
-                const auto tokStart = pos + flag.size();
-                auto tokEnd = tokStart;
-                while (tokEnd < cmd.size() && cmd[tokEnd] != L' ' && cmd[tokEnd] != L'\t')
-                {
-                    ++tokEnd;
-                }
-                if (tokEnd > tokStart)
-                {
-                    return cmd.substr(tokStart, tokEnd - tokStart);
-                }
-            }
-            // Commandline parse failed. Do NOT fall back to profile.Name() —
-            // that's user-editable display text (see
-            // doc/cascadia/profiles.schema.json), so using it as a distro
-            // identifier would write integration to the wrong distro or
-            // produce a wsl.exe spawn failure. Return empty; the caller
-            // (_SnapshotWslDistroNames) drops empty results.
-            return {};
-        }
-
-        // Snapshot the user's WSL distros from the live settings model.
-        //
-        // MUST be called on the UI thread — _settings.AllProfiles()
-        // returns the live observable vector, and iterating it
-        // concurrently with a settings reload (file watcher or user
-        // save on the UI thread) would be a use-after-mutate. We
-        // copy out plain strings so callers can move to a background
-        // thread to actually run the install / uninstall.
-        inline std::vector<std::wstring> _SnapshotWslDistroNames(
-            const winrt::Microsoft::Terminal::Settings::Model::CascadiaSettings& settings)
-        {
-            std::vector<std::wstring> out;
-            if (!settings)
-            {
-                return out;
-            }
-            // Multiple WT profiles can point at the same WSL distro
-            // (a user may duplicate "Ubuntu" with different startup
-            // commands, fonts, etc.). De-duplicate so each distro is
-            // touched exactly once per install/uninstall sweep — the
-            // per-distro path is also expensive (wsl.exe spawn, UNC
-            // mount, $HOME probe), so repeating it is doubly bad.
-            std::set<std::wstring> seen;
-            for (const auto& profile : settings.AllProfiles())
-            {
-                if (profile.Source() == L"Windows.Terminal.Wsl")
-                {
-                    auto name = _ExtractWslDistroName(profile);
-                    if (name.empty())
-                    {
-                        // _ExtractWslDistroName couldn't parse the
-                        // commandline. Skipping rather than attempting
-                        // an InstallWslBash("") that's guaranteed to
-                        // fail (and produces a noisy error in the FRE
-                        // / Settings UI). User's malformed WSL profile
-                        // simply doesn't get integration; the rest do.
-                        continue;
-                    }
-                    if (seen.insert(name).second)
-                    {
-                        out.emplace_back(std::move(name));
-                    }
-                }
-            }
-            return out;
-        }
-
-        // Snapshot which non-WSL shells the user has at least one
-        // profile for. Same UI-thread restriction as
-        // _SnapshotWslDistroNames — _settings.AllProfiles() is an
-        // observable live vector. Used to gate Install so we don't
-        // pollute HOME with a block for a shell the user has no
-        // profile for.
-        struct ShellPresence
-        {
-            bool pwsh{ false };
-            bool windowsPowerShell{ false };
-            bool bash{ false };
-        };
-        inline ShellPresence _SnapshotShellPresence(
-            const winrt::Microsoft::Terminal::Settings::Model::CascadiaSettings& settings)
-        {
-            ShellPresence out{};
-            if (!settings)
-            {
-                return out;
-            }
-            namespace SI = ::Microsoft::Terminal::ShellIntegration;
-            for (const auto& profile : settings.AllProfiles())
-            {
-                // Hold the hstring values in named locals so their lifetime
-                // extends through the wstring_view uses below — wstring_view
-                // is non-owning and a temporary hstring would dangle.
-                const auto src = profile.Source();
-                const auto cmd = profile.Commandline();
-                const std::wstring_view srcSv{ src };
-                const std::wstring_view cmdSv{ cmd };
-                if (!out.pwsh && SI::ProfileMatchesShell(SI::Target::Pwsh, srcSv, cmdSv))
-                {
-                    out.pwsh = true;
-                }
-                if (!out.windowsPowerShell && SI::ProfileMatchesShell(SI::Target::WindowsPowerShell, srcSv, cmdSv))
-                {
-                    out.windowsPowerShell = true;
-                }
-                if (!out.bash && SI::ProfileMatchesShell(SI::Target::Bash, srcSv, cmdSv))
-                {
-                    out.bash = true;
-                }
-                if (out.pwsh && out.windowsPowerShell && out.bash)
-                {
-                    break;
-                }
-            }
-            return out;
-        }
-    }
-
     FreOverlay::FreOverlay()
     {
         InitializeComponent();
@@ -986,40 +839,21 @@ namespace winrt::TerminalApp::implementation
             // thread BEFORE resuming on a background thread —
             // _settings.AllProfiles() is an observable vector and
             // iterating it concurrently with a settings reload is unsafe.
-            const auto wslDistros = _SnapshotWslDistroNames(_settings);
-            const auto shellPresence = _SnapshotShellPresence(_settings);
+            const auto wslDistros = ShellIntegrationSweep::SnapshotWslDistroNames(_settings);
+            const auto shellPresence = ShellIntegrationSweep::SnapshotShellPresence(_settings);
 
             co_await winrt::resume_background();
-            namespace SI = ::Microsoft::Terminal::ShellIntegration;
             // Profile-gated install: a user keeping only "Developer
             // PowerShell for VS" (Windows PowerShell) and no pwsh
             // profile must not get a pwsh integration block written.
-            // A skipped shell is reported as success-already-installed
-            // so the FRE failure verdict (below) doesn't flag a
-            // missing shell as an integration failure.
-            const auto skipped = []() {
-                SI::InstallResult r{};
-                r.success = true;
-                r.alreadyInstalled = true;
-                return r;
-            };
-            const auto pwsh7Result = shellPresence.pwsh
-                                         ? SI::InstallForTarget(SI::Target::Pwsh)
-                                         : skipped();
-            const auto windowsPsResult = shellPresence.windowsPowerShell
-                                             ? SI::InstallForTarget(SI::Target::WindowsPowerShell)
-                                             : skipped();
-            const auto bashResult = shellPresence.bash
-                                        ? SI::InstallForTarget(SI::Target::Bash)
-                                        : skipped();
-
-            // WSL distros — one install per distro the user has a WT
-            // profile for (covered by the snapshot above).
-            std::vector<std::pair<std::wstring, SI::InstallResult>> wslResults;
-            for (const auto& distName : wslDistros)
-            {
-                wslResults.emplace_back(distName, SI::InstallWslBash(distName));
-            }
+            // RunInstall reports a skipped shell as
+            // success-already-installed so the FRE failure verdict
+            // (below) doesn't flag a missing shell as a failure.
+            const auto results = ShellIntegrationSweep::RunInstall(shellPresence, wslDistros);
+            const auto& pwsh7Result = results.pwsh;
+            const auto& windowsPsResult = results.windowsPowerShell;
+            const auto& bashResult = results.bash;
+            const auto& wslResults = results.wsl;
 
             {
                 std::string detail = "[FRE] Shell integration: pwsh7=";
