+++ src/cascadia/TerminalApp/ShellIntegrationSweep.h
@@ -1,0 +1,226 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+//
+// ShellIntegrationSweep.h
+//
+// Shared profile snapshot + install/uninstall sweep used by all three
+// shell-integration entry points:
+//   • FreOverlay::Save           — FRE wizard "Install" button
+//   • TerminalPage::_InitShellIntegration — Settings UI "Install" button
+//   • TerminalPage::_ReconcileShellIntegration — startup + settings reload
+//
+// All three need the same two-phase pattern:
+//
+//   1. Snapshot the live `_settings.AllProfiles()` ON THE UI THREAD
+//      (the observable vector races with settings reload). The
+//      snapshot is two cheap copies:
+//        - WSL distro names (deduped, malformed entries dropped)
+//        - non-WSL ShellPresence bitset (pwsh / WinPS / Git Bash)
+//
+//   2. Run the install OR uninstall sweep on a background thread,
+//      using ONLY the snapshot — never re-touch settings.AllProfiles()
+//      from background code.
+//
+// Install AND uninstall are BOTH profile-gated: we only touch shells
+// the user has a profile for, and only WSL distros the user has a WT
+// profile for. Symmetry rationale: if a user toggles auto-detection
+// off, they expect cleanup for the shells they actually use. A
+// stale block for a shell the user never had a profile for would
+// stay (already true — we never wrote it), and the cost of touching
+// every shell on every off-toggle is asymmetric (writes a `.bak.*`
+// even when there's nothing to uninstall).
+//
+// CAVEAT: if the user installs for shell X, deletes the X profile,
+// then toggles off — the X block in their HOME survives. This matches
+// the install-time policy (profile presence is the gate), and the
+// next reconcile after re-adding the X profile will sweep it.
+
+#pragma once
+
+#include <set>
+#include <string>
+#include <vector>
+
+#include "../inc/ShellIntegration.h"
+#include "../inc/ShellIntegrationProfileGate.h"
+
+#include <winrt/Microsoft.Terminal.Settings.Model.h>
+
+namespace winrt::TerminalApp::implementation::ShellIntegrationSweep
+{
+    namespace SI = ::Microsoft::Terminal::ShellIntegration;
+    using CascadiaSettings = ::winrt::Microsoft::Terminal::Settings::Model::CascadiaSettings;
+    using Profile = ::winrt::Microsoft::Terminal::Settings::Model::Profile;
+
+    // Bitset of "user has at least one profile for this shell".
+    // WSL is NOT included — it has its own per-distro snapshot below.
+    struct ShellPresence
+    {
+        bool pwsh{ false };
+        bool windowsPowerShell{ false };
+        bool bash{ false };
+    };
+
+    // Parse a WSL distro name out of a profile generated by
+    // WslDistroGenerator. Prefer the `-d <distName>` / `--distribution
+    // <distName>` token in the commandline (the value WSL itself
+    // consumes) over profile.Name(): Name() is user-editable display
+    // text and would write integration to the wrong distro.
+    //
+    // Returns empty on parse failure (fail-closed); caller drops empties.
+    inline std::wstring ExtractWslDistroName(const Profile& profile)
+    {
+        const std::wstring cmd{ profile.Commandline() };
+        constexpr std::wstring_view shortFlag{ L" -d " };
+        constexpr std::wstring_view longFlag{ L" --distribution " };
+        for (const auto flag : { shortFlag, longFlag })
+        {
+            const auto pos = cmd.find(flag);
+            if (pos == std::wstring::npos)
+            {
+                continue;
+            }
+            const auto tokStart = pos + flag.size();
+            auto tokEnd = tokStart;
+            while (tokEnd < cmd.size() && cmd[tokEnd] != L' ' && cmd[tokEnd] != L'\t')
+            {
+                ++tokEnd;
+            }
+            if (tokEnd > tokStart)
+            {
+                return cmd.substr(tokStart, tokEnd - tokStart);
+            }
+        }
+        return {};
+    }
+
+    // Snapshot the user's distinct WSL distro names. MUST be called on
+    // the UI thread — settings.AllProfiles() is an observable vector
+    // and iterating concurrently with a reload is unsafe. Deduped
+    // (multiple WT profiles for the same distro touch it once).
+    // Empty / malformed extracts are dropped.
+    inline std::vector<std::wstring> SnapshotWslDistroNames(const CascadiaSettings& settings)
+    {
+        std::vector<std::wstring> out;
+        if (!settings)
+        {
+            return out;
+        }
+        std::set<std::wstring> seen;
+        for (const auto& profile : settings.AllProfiles())
+        {
+            if (profile.Source() == L"Windows.Terminal.Wsl")
+            {
+                auto name = ExtractWslDistroName(profile);
+                if (name.empty())
+                {
+                    continue;
+                }
+                if (seen.insert(name).second)
+                {
+                    out.emplace_back(std::move(name));
+                }
+            }
+        }
+        return out;
+    }
+
+    // Snapshot which non-WSL shells the user has a profile for. Same
+    // UI-thread restriction as SnapshotWslDistroNames.
+    inline ShellPresence SnapshotShellPresence(const CascadiaSettings& settings)
+    {
+        ShellPresence out{};
+        if (!settings)
+        {
+            return out;
+        }
+        for (const auto& profile : settings.AllProfiles())
+        {
+            const auto src = profile.Source();
+            const auto cmd = profile.Commandline();
+            const std::wstring_view srcSv{ src };
+            const std::wstring_view cmdSv{ cmd };
+            if (!out.pwsh && SI::ProfileMatchesShell(SI::Target::Pwsh, srcSv, cmdSv))
+            {
+                out.pwsh = true;
+            }
+            if (!out.windowsPowerShell && SI::ProfileMatchesShell(SI::Target::WindowsPowerShell, srcSv, cmdSv))
+            {
+                out.windowsPowerShell = true;
+            }
+            if (!out.bash && SI::ProfileMatchesShell(SI::Target::Bash, srcSv, cmdSv))
+            {
+                out.bash = true;
+            }
+            if (out.pwsh && out.windowsPowerShell && out.bash)
+            {
+                break;
+            }
+        }
+        return out;
+    }
+
+    // Result aggregate for the install sweep, surfaced to the FRE /
+    // Settings UI for the all-installed / any-failure verdict. WSL +
+    // Bash failures intentionally don't tank the verdict (a user
+    // without Git Bash / running WSL shouldn't see a false-alarm).
+    struct InstallSweepResults
+    {
+        SI::InstallResult pwsh{ true, true, {}, false };       // skipped → already-installed
+        SI::InstallResult windowsPowerShell{ true, true, {}, false };
+        SI::InstallResult bash{ true, true, {}, false };
+        std::vector<std::pair<std::wstring, SI::InstallResult>> wsl;
+    };
+
+    // Run the install sweep using the provided snapshot. Touches only
+    // shells the user has a profile for; touches each WSL distro once.
+    // Synchronous — call from a background thread.
+    inline InstallSweepResults RunInstall(const ShellPresence& shellPresence,
+                                          const std::vector<std::wstring>& wslDistros)
+    {
+        InstallSweepResults r{};
+        if (shellPresence.pwsh)
+        {
+            r.pwsh = SI::InstallForTarget(SI::Target::Pwsh);
+        }
+        if (shellPresence.windowsPowerShell)
+        {
+            r.windowsPowerShell = SI::InstallForTarget(SI::Target::WindowsPowerShell);
+        }
+        if (shellPresence.bash)
+        {
+            r.bash = SI::InstallForTarget(SI::Target::Bash);
+        }
+        r.wsl.reserve(wslDistros.size());
+        for (const auto& distName : wslDistros)
+        {
+            r.wsl.emplace_back(distName, SI::InstallWslBash(distName));
+        }
+        return r;
+    }
+
+    // Run the uninstall sweep using the provided snapshot. Mirrors
+    // RunInstall: profile-gated symmetric cleanup. A shell with no
+    // profile is left untouched (we never installed for it either).
+    // Synchronous — call from a background thread.
+    inline void RunUninstall(const ShellPresence& shellPresence,
+                             const std::vector<std::wstring>& wslDistros)
+    {
+        if (shellPresence.pwsh)
+        {
+            (void)SI::UninstallForTarget(SI::Target::Pwsh);
+        }
+        if (shellPresence.windowsPowerShell)
+        {
+            (void)SI::UninstallForTarget(SI::Target::WindowsPowerShell);
+        }
+        if (shellPresence.bash)
+        {
+            (void)SI::UninstallForTarget(SI::Target::Bash);
+        }
+        for (const auto& distName : wslDistros)
+        {
+            (void)SI::UninstallWslBash(distName);
+        }
+    }
+}
